pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "compass-rag"
version = "0.0.1"
description = "Vectorless Retrieval-Augmented Generation agent for OpenText Exstream documentation"
readme = "README.md"
requires-python = ">=3.11.9"
license = {text = "MIT"}
authors = [
    {name = "Exstream Team", email = "saikiran.aravala@gmail.com"}
]

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "langgraph>=0.1.0",
    "langchain>=0.2.0",
    "langchain-core>=0.2.0",
    "selectolax>=0.3.16",
    "readability-lxml>=0.8.1",
    "pypdf>=4.1.0",
    "pdfplumber>=0.11.0",
    "pytesseract>=0.3.10",
    "tantivy>=0.20.0",
    "openrouter>=0.1.0",
    "anthropic>=0.28.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "aiohttp>=3.9.1",
    "authlib>=1.3.0",
    "PyJWT>=2.8.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=0.41b0",
    "opentelemetry-exporter-jaeger-thrift>=1.20.0",
    "opentelemetry-exporter-prometheus>=0.41b0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-instrumentation-requests>=0.41b0",
    "opentelemetry-instrumentation-httpx>=0.41b0",
    "opentelemetry-instrumentation-logging>=0.41b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.41b0",
    "prometheus-client>=0.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "black==23.12.0",
    "ruff==0.1.8",
    "mypy==1.7.1",
    "ipython==8.17.2",
]

[project.urls]
Homepage = "https://github.com/anthropics/compass-rag"
Repository = "https://github.com/anthropics/compass-rag"
Issues = "https://github.com/anthropics/compass-rag/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=html"
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
